Move crypto bits into a subdirectory
It was getting a bit sprawly; this should help keep things together.
