+++ lib/crypto/OlmDevice.js
@@ -18,11 +18,11 @@ limitations under the License.
 /**
  * olm.js wrapper
  *
- * @module OlmDevice
+ * @module crypto/OlmDevice
  */
 
 var Olm = require("olm");
-var utils = require("./utils");
+var utils = require("../utils");
 
 /**
  * Manages the olm cryptography functions. Each OlmDevice has a single
@@ -31,7 +31,7 @@ var utils = require("./utils");
  * Accounts and sessions are kept pickled in a sessionStore.
  *
  * @constructor
- * @alias module:OlmDevice
+ * @alias module:crypto/OlmDevice
  *
  * @param {Object} sessionStore A store to be used for data in end-to-end
  *    crypto
