+++ lib/crypto/algorithms/base.js
@@ -18,23 +18,23 @@ limitations under the License.
 /**
  * Internal module. Defines the base classes of the encryption implementations
  *
- * @module crypto-algorithms/base
+ * @module crypto/algorithms/base
  */
-var utils = require("../utils");
+var utils = require("../../utils");
 
 /**
  * map of registered encryption algorithm classes. A map from string to {@link
- * module:crypto-algorithms/base.EncryptionAlgorithm|EncryptionAlgorithm} class
+ * module:crypto/algorithms/base.EncryptionAlgorithm|EncryptionAlgorithm} class
  *
- * @type {Object.<string, function(new: module:crypto-algorithms/base.EncryptionAlgorithm)>}
+ * @type {Object.<string, function(new: module:crypto/algorithms/base.EncryptionAlgorithm)>}
  */
 module.exports.ENCRYPTION_CLASSES = {};
 
 /**
  * map of registered encryption algorithm classes. Map from string to {@link
- * module:crypto-algorithms/base.DecryptionAlgorithm|DecryptionAlgorithm} class
+ * module:crypto/algorithms/base.DecryptionAlgorithm|DecryptionAlgorithm} class
  *
- * @type {Object.<string, function(new: module:crypto-algorithms/base.DecryptionAlgorithm)>}
+ * @type {Object.<string, function(new: module:crypto/algorithms/base.DecryptionAlgorithm)>}
  */
 module.exports.DECRYPTION_CLASSES = {};
 
@@ -42,12 +42,12 @@ module.exports.DECRYPTION_CLASSES = {};
  * base type for encryption implementations
  *
  * @constructor
- * @alias module:crypto-algorithms/base.EncryptionAlgorithm
+ * @alias module:crypto/algorithms/base.EncryptionAlgorithm
  *
  * @param {object} params parameters
  * @param {string} params.deviceId The identifier for this device.
  * @param {module:crypto} params.crypto crypto core
- * @param {module:OlmDevice} params.olmDevice olm.js wrapper
+ * @param {module:crypto/OlmDevice} params.olmDevice olm.js wrapper
  * @param {module:base-apis~MatrixBaseApis} baseApis base matrix api interface
  * @param {string} params.roomId  The ID of the room we will be sending to
  */
@@ -64,7 +64,7 @@ module.exports.EncryptionAlgorithm = EncryptionAlgorithm;
 /**
  * Encrypt a message event
  *
- * @method module:crypto-algorithms/base.EncryptionAlgorithm#encryptMessage
+ * @method module:crypto/algorithms/base.EncryptionAlgorithm#encryptMessage
  * @abstract
  *
  * @param {module:models/room} room
@@ -86,10 +86,10 @@ EncryptionAlgorithm.prototype.onRoomMembership = function(event, member) {};
  * base type for decryption implementations
  *
  * @constructor
- * @alias module:crypto-algorithms/base.DecryptionAlgorithm
+ * @alias module:crypto/algorithms/base.DecryptionAlgorithm
  *
  * @param {object} params parameters
- * @param {module:OlmDevice} params.olmDevice olm.js wrapper
+ * @param {module:crypto/OlmDevice} params.olmDevice olm.js wrapper
  */
 var DecryptionAlgorithm = function(params) {
     this._olmDevice = params.olmDevice;
@@ -100,21 +100,21 @@ module.exports.DecryptionAlgorithm = DecryptionAlgorithm;
 /**
  * Decrypt an event
  *
- * @method module:crypto-algorithms/base.DecryptionAlgorithm#decryptEvent
+ * @method module:crypto/algorithms/base.DecryptionAlgorithm#decryptEvent
  * @abstract
  *
  * @param {object} event raw event
  *
  * @return {object} decrypted payload (with properties 'type', 'content')
  *
- * @throws {module:crypto-algorithms/base.DecryptionError} if there is a
+ * @throws {module:crypto/algorithms/base.DecryptionError} if there is a
  *   problem decrypting the event
  */
 
 /**
  * Handle a key event
  *
- * @method module:crypto-algorithms/base.DecryptionAlgorithm#onRoomKeyEvent
+ * @method module:crypto/algorithms/base.DecryptionAlgorithm#onRoomKeyEvent
  *
  * @param {module:models/event.MatrixEvent} event key event
  */
@@ -140,11 +140,11 @@ utils.inherits(module.exports.DecryptionError, Error);
  * @param {string} algorithm algorithm tag to register for
  *
  * @param {class} encryptor {@link
- *     module:crypto-algorithms/base.EncryptionAlgorithm|EncryptionAlgorithm}
+ *     module:crypto/algorithms/base.EncryptionAlgorithm|EncryptionAlgorithm}
  *     implementation
  *
  * @param {class} decryptor {@link
- *     module:crypto-algorithms/base.DecryptionAlgorithm|DecryptionAlgorithm}
+ *     module:crypto/algorithms/base.DecryptionAlgorithm|DecryptionAlgorithm}
  *     implementation
  */
 module.exports.registerAlgorithm = function(algorithm, encryptor, decryptor) {
