+++ lib/crypto/algorithms/index.js
@@ -16,7 +16,7 @@ limitations under the License.
 "use strict";
 
 /**
- * @module crypto-algorithms
+ * @module crypto/algorithms
  */
 
 var base = require("./base");
@@ -25,16 +25,16 @@ require("./olm");
 require("./megolm");
 
 /**
- * @see module:crypto-algorithms/base.ENCRYPTION_CLASSES
+ * @see module:crypto/algorithms/base.ENCRYPTION_CLASSES
  */
 module.exports.ENCRYPTION_CLASSES = base.ENCRYPTION_CLASSES;
 
 /**
- * @see module:crypto-algorithms/base.DECRYPTION_CLASSES
+ * @see module:crypto/algorithms/base.DECRYPTION_CLASSES
  */
 module.exports.DECRYPTION_CLASSES = base.DECRYPTION_CLASSES;
 
 /**
- * @see module:crypto-algorithms/base.DecryptionError
+ * @see module:crypto/algorithms/base.DecryptionError
  */
 module.exports.DecryptionError = base.DecryptionError;
