+++ lib/crypto/algorithms/megolm.js
@@ -18,12 +18,12 @@ limitations under the License.
 /**
  * Defines m.olm encryption/decryption
  *
- * @module crypto-algorithms/megolm
+ * @module crypto/algorithms/megolm
  */
 
 var q = require("q");
 
-var utils = require("../utils");
+var utils = require("../../utils");
 var olmlib = require("../olmlib");
 var base = require("./base");
 
@@ -31,10 +31,10 @@ var base = require("./base");
  * Megolm encryption implementation
  *
  * @constructor
- * @extends {module:crypto-algorithms/base.EncryptionAlgorithm}
+ * @extends {module:crypto/algorithms/base.EncryptionAlgorithm}
  *
  * @param {object} params parameters, as per
- *     {@link module:crypto-algorithms/base.EncryptionAlgorithm}
+ *     {@link module:crypto/algorithms/base.EncryptionAlgorithm}
  */
 function MegolmEncryption(params) {
     base.EncryptionAlgorithm.call(this, params);
@@ -211,10 +211,10 @@ MegolmEncryption.prototype.onRoomMembership = function(event, member) {
  * Megolm decryption implementation
  *
  * @constructor
- * @extends {module:crypto-algorithms/base.DecryptionAlgorithm}
+ * @extends {module:crypto/algorithms/base.DecryptionAlgorithm}
  *
  * @param {object} params parameters, as per
- *     {@link module:crypto-algorithms/base.DecryptionAlgorithm}
+ *     {@link module:crypto/algorithms/base.DecryptionAlgorithm}
  */
 function MegolmDecryption(params) {
     base.DecryptionAlgorithm.call(this, params);
@@ -228,7 +228,7 @@ utils.inherits(MegolmDecryption, base.DecryptionAlgorithm);
  *
  * @return {object} decrypted payload (with properties 'type', 'content')
  *
- * @throws {module:crypto-algorithms/base.DecryptionError} if there is a
+ * @throws {module:crypto/algorithms/base.DecryptionError} if there is a
  *   problem decrypting the event
  */
 MegolmDecryption.prototype.decryptEvent = function(event) {
