+++ lib/crypto/algorithms/olm.js
@@ -18,13 +18,13 @@ limitations under the License.
 /**
  * Defines m.olm encryption/decryption
  *
- * @module crypto-algorithms/olm
+ * @module crypto/algorithms/olm
  */
 var q = require('q');
 
-var utils = require("../utils");
+var utils = require("../../utils");
 var olmlib = require("../olmlib");
-var DeviceInfo = require("../crypto-deviceinfo");
+var DeviceInfo = require("../deviceinfo");
 var DeviceVerification = DeviceInfo.DeviceVerification;
 
 
@@ -34,10 +34,10 @@ var base = require("./base");
  * Olm encryption implementation
  *
  * @constructor
- * @extends {module:crypto-algorithms/base.EncryptionAlgorithm}
+ * @extends {module:crypto/algorithms/base.EncryptionAlgorithm}
  *
  * @param {object} params parameters, as per
- *     {@link module:crypto-algorithms/base.EncryptionAlgorithm}
+ *     {@link module:crypto/algorithms/base.EncryptionAlgorithm}
  */
 function OlmEncryption(params) {
     base.EncryptionAlgorithm.call(this, params);
@@ -128,9 +128,9 @@ OlmEncryption.prototype.encryptMessage = function(room, eventType, content) {
  * Olm decryption implementation
  *
  * @constructor
- * @extends {module:crypto-algorithms/base.DecryptionAlgorithm}
+ * @extends {module:crypto/algorithms/base.DecryptionAlgorithm}
  * @param {object} params parameters, as per
- *     {@link module:crypto-algorithms/base.DecryptionAlgorithm}
+ *     {@link module:crypto/algorithms/base.DecryptionAlgorithm}
  */
 function OlmDecryption(params) {
     base.DecryptionAlgorithm.call(this, params);
@@ -144,7 +144,7 @@ utils.inherits(OlmDecryption, base.DecryptionAlgorithm);
  *
  * @return {object} decrypted payload (with properties 'type', 'content')
  *
- * @throws {module:crypto-algorithms/base.DecryptionError} if there is a
+ * @throws {module:crypto/algorithms/base.DecryptionError} if there is a
  *   problem decrypting the event
  */
 OlmDecryption.prototype.decryptEvent = function(event) {
