+++ lib/crypto/deviceinfo.js
@@ -17,14 +17,14 @@ limitations under the License.
 
 
 /**
- * @module crypto-deviceinfo
+ * @module crypto/deviceinfo
  */
 
 /**
   * Information about a user's device
   *
   * @constructor
-  * @alias module:crypto-deviceinfo
+  * @alias module:crypto/deviceinfo
   *
   * @property {string} deviceId the ID of this device
   *
@@ -33,7 +33,7 @@ limitations under the License.
   * @property {Object.<string,string>} keys a map from
   *      &lt;key type&gt;:&lt;id&gt; -> &lt;base64-encoded key&gt;>
   *
-  * @property {module:crypto-deviceinfo.DeviceVerification} verified
+  * @property {module:crypto/deviceinfo.DeviceVerification} verified
   *     whether the device has been verified by the user
   *
   * @property {Object} unsigned  additional data from the homeserver
