+++ lib/crypto/index.js
@@ -23,17 +23,16 @@ limitations under the License.
 var anotherjson = require('another-json');
 var q = require("q");
 
-var utils = require("./utils");
+var utils = require("../utils");
 var OlmDevice = require("./OlmDevice");
 var olmlib = require("./olmlib");
-var algorithms = require("./crypto-algorithms");
-var DeviceInfo = require("./crypto-deviceinfo");
+var algorithms = require("./algorithms");
+var DeviceInfo = require("./deviceinfo");
 var DeviceVerification = DeviceInfo.DeviceVerification;
 
 /**
  * Cryptography bits
  *
- * @alias module:crypto.Crypto
  * @constructor
  * @alias module:crypto
  *
@@ -198,7 +197,7 @@ function _uploadOneTimeKeys(crypto) {
  * @param {bool} forceDownload Always download the keys even if cached.
  *
  * @return {Promise} A promise which resolves to a map userId->deviceId->{@link
- * module:crypto-deviceinfo|DeviceInfo}.
+ * module:crypto/deviceinfo|DeviceInfo}.
  */
 Crypto.prototype.downloadKeys = function(userIds, forceDownload) {
     var self = this;
@@ -371,7 +370,7 @@ function _storeDeviceKeys(_olmDevice, userId, deviceId, userStore, deviceResult)
  *
  * @param {string} userId the user to list keys for.
  *
- * @return {module:crypto-deviceinfo[]} list of devices
+ * @return {module:crypto/deviceinfo[]} list of devices
  */
 Crypto.prototype.getStoredDevicesForUser = function(userId) {
     var devs = this._sessionStore.getEndToEndDevicesForUser(userId);
@@ -434,7 +433,7 @@ Crypto.prototype.listDeviceKeys = function(userId) {
  * @param {string} algorithm  encryption algorithm
  * @param {string} sender_key curve25519 key to match
  *
- * @return {module:crypto-deviceinfo?}
+ * @return {module:crypto/deviceinfo?}
  */
 Crypto.prototype.getDeviceByIdentityKey = function(userId, algorithm, sender_key) {
     if (
@@ -521,7 +520,8 @@ Crypto.prototype.setDeviceVerification = function(userId, deviceId, verified, bl
  * <p>
  * Returns a map from device id to an object with keys 'deviceIdKey' (the
  * device's curve25519 identity key) and 'sessions' (an array of objects in the
- * same format as that returned by {@link module:OlmDevice#getSessionInfoForDevice}).
+ * same format as that returned by
+ * {@link module:crypto/OlmDevice#getSessionInfoForDevice}).
  * <p>
  * This method is provided for debugging purposes.
  *
@@ -619,7 +619,7 @@ Crypto.prototype.setRoomEncryption = function(roomId, config) {
 
 /**
  * @typedef {Object} module:crypto~OlmSessionResult
- * @property {module:crypto-deviceinfo} device  device info
+ * @property {module:crypto/deviceinfo} device  device info
  * @property {string?} sessionId base64 olm session id; null if no session
  *    could be established
  */
@@ -863,7 +863,7 @@ Crypto.prototype._onRoomMembership = function(event, member) {
 };
 
 /**
- * @see module:crypto-algorithms/base.DecryptionError
+ * @see module:crypto/algorithms/base.DecryptionError
  */
 Crypto.DecryptionError = algorithms.DecryptionError;
