+++ lib/crypto/olmlib.js
@@ -20,7 +20,7 @@ limitations under the License.
  * Utilities common to olm encryption algorithms
  */
 
-var utils = require("./utils");
+var utils = require("../utils");
 
 /**
  * matrix algorithm tag for olm
@@ -37,7 +37,7 @@ module.exports.MEGOLM_ALGORITHM = "m.megolm.v1.aes-sha2";
  * Encrypt an event payload for a list of devices
  *
  * @param {string} ourDeviceId
- * @param {module:OlmDevice} olmDevice olm.js wrapper
+ * @param {module:crypto/OlmDevice} olmDevice olm.js wrapper
  * @param {string[]} participantKeys list of curve25519 keys to encrypt for
  * @param {object} payloadFields fields to include in the encrypted payload
  *
